test(actions): consolidate confirm/successOperations tests into actions-popover
Per testing standards: test user-visible behavior at the highest realistic
level, no duplication across unit and integration layers.

- Delete confirm-dispatcher.test.tsx — all confirm flow behavior (cancel,
  successOperations toast/failure) now tested end-to-end through ActionsPopover
- Trim use-success-operations.test.tsx to the 3 invalidate tests only
  (queryKey shape, serviceOptions, ordering) — these assert internal decisions
  invisible at the DOM level; toast/interpolation behavior moved up
- Add 7 new integration tests to actions-popover.test.tsx: cancel, toast on
  success/failure, toast with/without route, response interpolation
- Add getSnackbarProviderWrapper to existing confirm tests (ConfirmDispatcher
  now calls useSnackbar via useSuccessOperations)
