+++ javascript/packages/core/components/actions/__tests__/actions-popover.test.tsx
@@ -3,6 +3,7 @@ import userEvent from '@testing-library/user-event';
 
 import { ActionMenu } from '#core/components/actions/action-menu/action-menu';
 import { ActionsPopover } from '#core/components/actions/actions-popover';
+import { interpolate } from '#core/interpolation/interpolate';
 import { buildWrapper } from '#core/test/wrappers/build-wrapper';
 import { getBaseProviderWrapper } from '#core/test/wrappers/get-base-provider-wrapper';
 import { getErrorProviderWrapper } from '#core/test/wrappers/get-error-provider-wrapper';
@@ -12,6 +13,7 @@ import {
   createQueryMockRouter,
   getServiceProviderWrapper,
 } from '#core/test/wrappers/get-service-provider-wrapper';
+import { getSnackbarProviderWrapper } from '#core/test/wrappers/get-snackbar-provider-wrapper';
 
 import type { ActionComponentProps } from '#core/components/actions/types';
 
@@ -430,6 +432,7 @@ describe('ActionsPopover', () => {
         getErrorProviderWrapper(),
         getRouterWrapper(),
         getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
       ])
     );
 
@@ -472,6 +475,7 @@ describe('ActionsPopover', () => {
         getErrorProviderWrapper(),
         getRouterWrapper(),
         getServiceProviderWrapper({ request: failingRequest }),
+        getSnackbarProviderWrapper(),
       ])
     );
 
@@ -508,6 +512,7 @@ describe('ActionsPopover', () => {
         getErrorProviderWrapper(),
         getRouterWrapper({ location: '/start' }),
         getServiceProviderWrapper({ request: vi.fn() }),
+        getSnackbarProviderWrapper(),
       ])
     );
 
@@ -517,5 +522,279 @@ describe('ActionsPopover', () => {
     await user.click(within(dialog).getByRole('button', { name: 'Open' }));
 
     expect(screen.getByText(/Current pathname: \/dest\/page/)).toBeInTheDocument();
+  });
+
+  it('mutation-confirm action: cancel closes dialog without mutating', async () => {
+    const user = userEvent.setup();
+    const mockRequest = createQueryMockRouter({ UpdateTriggerRun: {} });
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Kill' },
+            operation: { type: 'mutation', mutation: { mutationName: 'UpdateTriggerRun' } },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm kill?' },
+              button: { label: 'Kill it' },
+              destructive: true,
+            },
+          },
+        ]}
+        record={{ id: 'run-1' }}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper(),
+        getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Kill' }));
+    const dialog = await screen.findByRole('dialog', { name: 'Confirm kill?' });
+    await user.click(within(dialog).getByRole('button', { name: 'Cancel' }));
+
+    await waitFor(() => {
+      expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
+    });
+    expect(mockRequest).not.toHaveBeenCalled();
+  });
+
+  it('mutation-confirm action: shows success toast after mutation', async () => {
+    const user = userEvent.setup();
+    const mockRequest = createQueryMockRouter({ UpdateTriggerRun: {} });
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Kill' },
+            operation: {
+              type: 'mutation',
+              mutation: { mutationName: 'UpdateTriggerRun' },
+              successOperations: [{ type: 'toast', message: 'Trigger killed' }],
+            },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm kill?' },
+              button: { label: 'Kill it' },
+            },
+          },
+        ]}
+        record={{ id: 'run-1' }}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper(),
+        getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Kill' }));
+    await user.click(
+      within(await screen.findByRole('dialog', { name: 'Confirm kill?' })).getByRole('button', {
+        name: 'Kill it',
+      })
+    );
+
+    expect(await screen.findByText('Trigger killed')).toBeInTheDocument();
+  });
+
+  it('mutation-confirm action: does not show success toast on failure', async () => {
+    const user = userEvent.setup();
+    const failingRequest = vi.fn().mockRejectedValue(new Error('rpc error'));
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Kill' },
+            operation: {
+              type: 'mutation',
+              mutation: { mutationName: 'UpdateTriggerRun' },
+              successOperations: [{ type: 'toast', message: 'Should not appear' }],
+            },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm kill?' },
+              button: { label: 'Kill it' },
+            },
+          },
+        ]}
+        record={{ id: 'run-1' }}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper(),
+        getServiceProviderWrapper({ request: failingRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Kill' }));
+    const dialog = await screen.findByRole('dialog', { name: 'Confirm kill?' });
+    await user.click(within(dialog).getByRole('button', { name: 'Kill it' }));
+
+    await within(dialog).findByText(/Test error/);
+    expect(screen.queryByText('Should not appear')).not.toBeInTheDocument();
+  });
+
+  it('mutation-confirm action: success toast with route navigates on click', async () => {
+    const user = userEvent.setup();
+    const mockRequest = createQueryMockRouter({ UpdateTriggerRun: {} });
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Kill' },
+            operation: {
+              type: 'mutation',
+              mutation: { mutationName: 'UpdateTriggerRun' },
+              successOperations: [
+                {
+                  type: 'toast',
+                  message: 'Done',
+                  action: { label: 'View', route: '/detail/run-1' },
+                },
+              ],
+            },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm kill?' },
+              button: { label: 'Kill it' },
+            },
+          },
+        ]}
+        record={{ id: 'run-1' }}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper({ location: '/start' }),
+        getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Kill' }));
+    await user.click(
+      within(await screen.findByRole('dialog', { name: 'Confirm kill?' })).getByRole('button', {
+        name: 'Kill it',
+      })
+    );
+    await screen.findByText('Done');
+    await user.click(screen.getAllByRole('button', { name: 'View' })[0]);
+
+    expect(screen.getByText(/Current pathname: \/detail\/run-1/)).toBeInTheDocument();
+  });
+
+  it('mutation-confirm action: success toast without route dismisses on click', async () => {
+    const user = userEvent.setup();
+    const mockRequest = createQueryMockRouter({ UpdateTriggerRun: {} });
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Kill' },
+            operation: {
+              type: 'mutation',
+              mutation: { mutationName: 'UpdateTriggerRun' },
+              successOperations: [{ type: 'toast', message: 'Done', action: { label: 'OK' } }],
+            },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm kill?' },
+              button: { label: 'Kill it' },
+            },
+          },
+        ]}
+        record={{ id: 'run-1' }}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper({ location: '/start' }),
+        getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Kill' }));
+    await user.click(
+      within(await screen.findByRole('dialog', { name: 'Confirm kill?' })).getByRole('button', {
+        name: 'Kill it',
+      })
+    );
+    await screen.findByText('Done');
+    await user.click(screen.getAllByRole('button', { name: 'OK' })[0]);
+
+    await waitFor(() => {
+      expect(screen.queryByText('Done')).not.toBeInTheDocument();
+    });
+    expect(screen.getByText(/Current pathname: \/start/)).toBeInTheDocument();
+  });
+
+  it('mutation-confirm action: success toast interpolates mutation response', async () => {
+    const user = userEvent.setup();
+    const mockRequest = createQueryMockRouter({ CreatePipelineRun: { name: 'training-v2' } });
+
+    render(
+      <ActionsPopover
+        actions={[
+          {
+            display: { label: 'Create' },
+            operation: {
+              type: 'mutation',
+              mutation: { mutationName: 'CreatePipelineRun' },
+              successOperations: [
+                { type: 'toast', message: interpolate('Pipeline ${response.name} created') },
+              ],
+            },
+            modal: {
+              type: 'confirm',
+              header: { title: 'Confirm create?' },
+              button: { label: 'Create' },
+            },
+          },
+        ]}
+        record={{}}
+      />,
+      buildWrapper([
+        getBaseProviderWrapper(),
+        getIconProviderWrapper(),
+        getErrorProviderWrapper(),
+        getRouterWrapper(),
+        getServiceProviderWrapper({ request: mockRequest }),
+        getSnackbarProviderWrapper(),
+      ])
+    );
+
+    await user.click(screen.getByRole('button', { name: 'Actions' }));
+    await user.click(await screen.findByRole('option', { name: 'Create' }));
+    await user.click(
+      within(await screen.findByRole('dialog', { name: 'Confirm create?' })).getByRole('button', {
+        name: 'Create',
+      })
+    );
+
+    expect(await screen.findByText('Pipeline training-v2 created')).toBeInTheDocument();
   });
 });
